Split method validation generation to a single file
