+++ packages/@react-native-windows/codegen/src/generators/GenerateNM2.ts
@@ -6,15 +6,10 @@
 
 'use strict';
 
-import {
-  NativeModuleFunctionTypeAnnotation,
-  NativeModulePropertyShape,
-  SchemaType,
-} from 'react-native-tscodegen';
+import {SchemaType} from 'react-native-tscodegen';
 import {AliasMap, setPreferredModuleName} from './AliasManaging';
 import {createAliasMap, generateAliases} from './AliasGen';
-import {translateArgs, translateSpecArgs} from './ParamTypes';
-import {translateImplReturnType, translateSpecReturnType} from './ReturnTypes';
+import {generateValidateMethods} from './ValidateMethods';
 
 type FilesOutput = Map<string, string>;
 
@@ -34,129 +29,12 @@ const moduleTemplate = `
 namespace ::_NAMESPACE_:: {
 ::_MODULE_ALIASED_STRUCTS_::
 struct ::_MODULE_NAME_::Spec : winrt::Microsoft::ReactNative::TurboModuleSpec {
-  static constexpr auto methods = std::tuple{
-::_MODULE_PROPERTIES_TUPLE_::
-  };
-
-  template <class TModule>
-  static constexpr void ValidateModule() noexcept {
-    constexpr auto methodCheckResults = CheckMethods<TModule, ::_MODULE_NAME_::Spec>();
-
-::_MODULE_PROPERTIES_SPEC_ERRORS_::
+::_MODULE_MEMBERS_VALIDATION_::
   }
 };
 
 } // namespace ::_NAMESPACE_::
 `;
-
-function isMethodSync(funcType: NativeModuleFunctionTypeAnnotation) {
-  return (
-    funcType.returnTypeAnnotation.type !== 'VoidTypeAnnotation' &&
-    funcType.returnTypeAnnotation.type !== 'PromiseTypeAnnotation'
-  );
-}
-
-function isMethodReturnPromise(funcType: NativeModuleFunctionTypeAnnotation) {
-  return funcType.returnTypeAnnotation.type === 'PromiseTypeAnnotation';
-}
-
-function getPossibleMethodSignatures(
-  prop: NativeModulePropertyShape,
-  funcType: NativeModuleFunctionTypeAnnotation,
-  aliases: AliasMap,
-  baseAliasName: string,
-): string[] {
-  const args = translateArgs(funcType.params, aliases, baseAliasName);
-  if (isMethodReturnPromise(funcType)) {
-    // TODO: type of the promise could be provided in the future
-    args.push('React::ReactPromise<React::JSValue> &&result');
-  }
-
-  // TODO: be much more exhastive on the possible method signatures that can be used..
-  const sig = `REACT_${isMethodSync(funcType) ? 'SYNC_' : ''}METHOD(${
-    prop.name
-  }) ${translateImplReturnType(
-    funcType.returnTypeAnnotation,
-    aliases,
-    baseAliasName,
-  )} ${prop.name}(${args.join(', ')}) noexcept { /* implementation */ }}`;
-
-  const staticsig = `REACT_${isMethodSync(funcType) ? 'SYNC_' : ''}METHOD(${
-    prop.name
-  }) static ${translateImplReturnType(
-    funcType.returnTypeAnnotation,
-    aliases,
-    baseAliasName,
-  )} ${prop.name}(${args.join(', ')}) noexcept { /* implementation */ }}`;
-
-  return [sig, staticsig];
-}
-
-function translatePossibleMethodSignatures(
-  prop: NativeModulePropertyShape,
-  funcType: NativeModuleFunctionTypeAnnotation,
-  aliases: AliasMap,
-  baseAliasName: string,
-): string {
-  return getPossibleMethodSignatures(prop, funcType, aliases, baseAliasName)
-    .map(sig => `"    ${sig}\\n"`)
-    .join('\n          ');
-}
-
-function renderProperties(
-  properties: ReadonlyArray<NativeModulePropertyShape>,
-  aliases: AliasMap,
-  tuple: boolean,
-): string {
-  // TODO: generate code for constants
-  return properties
-    .filter(prop => prop.name !== 'getConstants')
-    .map((prop, index) => {
-      // TODO: prop.optional === true
-      // TODO: prop.typeAnnotation.type === 'NullableTypeAnnotation'
-      const propAliasName = prop.name;
-      const funcType =
-        prop.typeAnnotation.type === 'NullableTypeAnnotation'
-          ? prop.typeAnnotation.typeAnnotation
-          : prop.typeAnnotation;
-
-      const traversedArgs = translateSpecArgs(
-        funcType.params,
-        aliases,
-        propAliasName,
-      );
-
-      const translatedReturnParam = translateSpecReturnType(
-        funcType.returnTypeAnnotation,
-        aliases,
-        propAliasName,
-      );
-
-      if (isMethodReturnPromise(funcType)) {
-        // TODO: type of the promise could be provided in the future
-        traversedArgs.push('Promise<React::JSValue>');
-      }
-
-      if (tuple) {
-        return `      ${
-          isMethodSync(funcType) ? 'Sync' : ''
-        }Method<${translatedReturnParam}(${traversedArgs.join(
-          ', ',
-        )}) noexcept>{${index}, L"${prop.name}"},`;
-      } else {
-        return `    REACT_SHOW_METHOD_SPEC_ERRORS(
-          ${index},
-          "${prop.name}",
-          ${translatePossibleMethodSignatures(
-            prop,
-            funcType,
-            aliases,
-            propAliasName,
-          )});`;
-      }
-    })
-    .join('\n');
-}
 
 export function createNM2Generator({namespace}: {namespace: string}) {
   return (
@@ -182,18 +60,8 @@ export function createNM2Generator({namespace}: {namespace: string}) {
         // copy all explicit to a map
         const aliases: AliasMap = createAliasMap(nativeModule.aliases);
 
-        // prepare members for turbo modules
-        const properties = nativeModule.spec.properties;
-        const traversedProperties = renderProperties(
-          properties,
-          aliases,
-          false,
-        );
-        const traversedPropertyTuples = renderProperties(
-          properties,
-          aliases,
-          true,
-        );
+        // prepare methods
+        const methods = generateValidateMethods(nativeModule, aliases);
 
         // generate code for structs
         const traversedAliasedStructs = generateAliases(aliases);
@@ -202,11 +70,7 @@ export function createNM2Generator({namespace}: {namespace: string}) {
           `Native${preferredModuleName}Spec.g.h`,
           moduleTemplate
             .replace(/::_MODULE_ALIASED_STRUCTS_::/g, traversedAliasedStructs)
-            .replace(/::_MODULE_PROPERTIES_TUPLE_::/g, traversedPropertyTuples)
-            .replace(
-              /::_MODULE_PROPERTIES_SPEC_ERRORS_::/g,
-              traversedProperties,
-            )
+            .replace(/::_MODULE_MEMBERS_VALIDATION_::/g, methods)
             .replace(/::_MODULE_NAME_::/g, preferredModuleName)
             .replace(/::_NAMESPACE_::/g, namespace),
         );
